refactor(webview): split the alert rule dialog into composed parts
