+++ apps/webview-ui/src/features/alert/ui/components/alerts-settings.tsx
@@ -1,37 +1,18 @@
 'use client';
 
-import { zodResolver } from '@hookform/resolvers/zod';
-import type {
-  AlertIntegration,
-  AlertRule,
-  AlertType,
-  Project,
-} from '@rustrak/client';
-import {
-  Bell,
-  BellOff,
-  Hash,
-  Loader2,
-  Mail,
-  Minus,
-  Pencil,
-  Plus,
-  RefreshCw,
-  Trash2,
-  Webhook,
-  Zap,
-} from 'lucide-react';
+import type { AlertIntegration, AlertRule, Project } from '@rustrak/client';
+import { Bell, Pencil, Plus, Trash2 } from 'lucide-react';
 import { useRouter } from 'next/navigation';
 import { useState, useTransition } from 'react';
-import { useForm } from 'react-hook-form';
 import { toast } from 'sonner';
-import { z } from 'zod';
 import {
-  createAlertRule,
   deleteAlertRule,
   updateAlertRule,
 } from '@/features/alert/api/mutations';
-import { applyServerFieldErrors } from '@/shared/lib/form-errors';
+import { alertTypeInfo } from '@/features/alert/model/alert-types';
+import { AlertRuleFormDialog } from '@/features/alert/ui/components/alert-rule-dialog/alert-rule-dialog';
+import { AlertTypeIcon } from '@/features/alert/ui/components/alert-type-icon';
+import { ProviderIcon } from '@/features/alert/ui/components/provider-icon';
 import { cn } from '@/shared/lib/utils';
 import {
   AlertDialog,
@@ -45,25 +26,6 @@ import {
 } from '@/shared/ui/components/shadcn/alert-dialog';
 import { Badge } from '@/shared/ui/components/shadcn/badge';
 import { Button } from '@/shared/ui/components/shadcn/button';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/shared/ui/components/shadcn/dialog';
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-  FormRootError,
-} from '@/shared/ui/components/shadcn/form';
-import { Input } from '@/shared/ui/components/shadcn/input';
 import { Switch } from '@/shared/ui/components/shadcn/switch';
 import {
   Table,
@@ -73,120 +35,6 @@ import {
   TableHeader,
   TableRow,
 } from '@/shared/ui/components/shadcn/table';
-import { Textarea } from '@/shared/ui/components/shadcn/textarea';
-
-// Alert type definitions
-const alertTypes: {
-  type: AlertType;
-  name: string;
-  description: string;
-  icon: React.ElementType;
-}[] = [
-  {
-    type: 'new_issue',
-    name: 'New Issue',
-    description: 'A new issue is first detected',
-    icon: Zap,
-  },
-  {
-    type: 'regression',
-    name: 'Regression',
-    description: 'A resolved issue reappears',
-    icon: RefreshCw,
-  },
-  {
-    type: 'unmute',
-    name: 'Unmute',
-    description: 'A muted issue is unmuted',
-    icon: BellOff,
-  },
-];
-
-// Per-channel routing_override map keyed by integration ID
-type RoutingMap = Record<number, Record<string, string>>;
-
-// Form schema
-const alertRuleFormSchema = z.object({
-  name: z.string().min(1, 'Name is required').max(255),
-  alert_type: z.enum(['new_issue', 'regression', 'unmute']),
-  selected_integration_ids: z.array(z.number()),
-  routing_map: z
-    .record(z.string(), z.record(z.string(), z.string()))
-    .optional(),
-  is_enabled: z.boolean(),
-  cooldown_minutes: z.number().int().min(0),
-});
-
-type AlertRuleFormData = z.infer<typeof alertRuleFormSchema>;
-
-// Provider icon helper
-function ProviderIcon({
-  type,
-  className,
-}: {
-  type: string;
-  className?: string;
-}) {
-  switch (type) {
-    case 'slack':
-      return <Hash className={className} />;
-    case 'email':
-      return <Mail className={className} />;
-    case 'webhook':
-      return <Webhook className={className} />;
-    default:
-      return <Bell className={className} />;
-  }
-}
-
-function getSlackMethod(integration: AlertIntegration): string {
-  return (
-    ((integration.credentials as Record<string, unknown>).method as string) ??
-    'webhook'
-  );
-}
-
-function validateRoutingForIntegration(
-  integration: AlertIntegration,
-  routing: Record<string, string>,
-): string | null {
-  if (integration.provider_type === 'slack') {
-    if (getSlackMethod(integration) === 'bot_token') {
-      if (!routing.channel || routing.channel.trim() === '') {
-        return `Slack channel is required for "${integration.name}"`;
-      }
-    }
-  }
-  if (integration.provider_type === 'email') {
-    if (!routing.recipients || routing.recipients.trim() === '') {
-      return `Recipients are required for "${integration.name}"`;
-    }
-    const addrs = routing.recipients
-      .split(',')
-      .map((a) => a.trim())
-      .filter(Boolean);
-    if (addrs.length === 0 || addrs.some((a) => !a.includes('@'))) {
-      return `Invalid email address in recipients for "${integration.name}"`;
-    }
-  }
-  if (integration.provider_type === 'webhook') {
-    const credUrl = (integration.credentials as Record<string, unknown>).url as
-      | string
-      | undefined;
-    const routeUrl = routing.url?.trim();
-    if (!credUrl && !routeUrl) {
-      return `A webhook URL is required for "${integration.name}" (set in credentials or as override URL)`;
-    }
-    if (
-      routeUrl &&
-      !routeUrl.startsWith('http://') &&
-      !routeUrl.startsWith('https://')
-    ) {
-      return `Override URL for "${integration.name}" must be http/https`;
-    }
-  }
-  return null;
-}
 
 interface AlertsSettingsProps {
   project: Project;
@@ -206,14 +54,6 @@ export function AlertsSettings({
   const [showAddForm, setShowAddForm] = useState(false);
 
   const enabledIntegrations = channels.filter((c) => c.is_enabled);
-
-  const getAlertTypeInfo = (type: string) =>
-    alertTypes.find((t) => t.type === type) ?? {
-      type,
-      name: type,
-      description: '',
-      icon: Bell,
-    };
 
   const getIntegrationById = (id: number) => channels.find((c) => c.id === id);
 
@@ -311,8 +151,7 @@ export function AlertsSettings({
             </TableHeader>
             <TableBody>
               {alertRules.map((rule) => {
-                const typeInfo = getAlertTypeInfo(rule.alert_type);
-                const TypeIcon = typeInfo.icon;
+                const typeInfo = alertTypeInfo(rule.alert_type);
                 return (
                   <TableRow
                     key={rule.id}
@@ -321,7 +160,10 @@ export function AlertsSettings({
                     <TableCell className="font-medium">{rule.name}</TableCell>
                     <TableCell>
                       <span className="flex items-center gap-1.5 text-sm text-muted-foreground">
-                        <TypeIcon className="size-3.5" />
+                        <AlertTypeIcon
+                          type={rule.alert_type}
+                          className="size-3.5"
+                        />
                         {typeInfo.name}
                       </span>
                     </TableCell>
@@ -439,630 +281,6 @@ export function AlertsSettings({
           </AlertDialogFooter>
         </AlertDialogContent>
       </AlertDialog>
-    </>
-  );
-}
-
-// ============================================================================
-// Alert Rule Form Dialog
-// ============================================================================
-
-interface AlertRuleFormDialogProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  projectId: number;
-  integrations: AlertIntegration[];
-  existingRule: AlertRule | null;
-  existingRuleTypes: string[];
-  onSuccess: () => void;
-}
-
-/**
- * The routing overrides an existing rule already carries, flattened into the
- * one-string-per-field shape the inputs bind to. Email is special-cased
- * because its `recipients` is a list on the wire and a comma-joined line in
- * the UI.
- */
-function initialRoutingMap(
-  existingRule: AlertRule | null,
-  integrations: AlertIntegration[],
-): RoutingMap {
-  if (!existingRule) return {};
-  const map: RoutingMap = {};
-  for (const ch of existingRule.channels) {
-    const override = ch.routing_override ?? {};
-    const integration = integrations.find((i) => i.id === ch.integration_id);
-    const normalized: Record<string, string> = {};
-    if (integration?.provider_type === 'email') {
-      const r = override.recipients;
-      if (Array.isArray(r)) normalized.recipients = r.join(', ');
-    } else {
-      for (const [k, v] of Object.entries(override)) {
-        if (typeof v === 'string') normalized[k] = v;
-      }
-    }
-    map[ch.integration_id] = normalized;
-  }
-  return map;
-}
-
-function AlertRuleFormDialog({
-  open,
-  onOpenChange,
-  ...bodyProps
-}: AlertRuleFormDialogProps) {
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="sm:max-w-lg max-h-[90vh] overflow-y-auto">
-        <AlertRuleForm
-          key={bodyProps.existingRule?.id ?? 'new'}
-          onOpenChange={onOpenChange}
-          {...bodyProps}
-        />
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function AlertRuleForm({
-  onOpenChange,
-  projectId,
-  integrations,
-  existingRule,
-  existingRuleTypes,
-  onSuccess,
-}: Omit<AlertRuleFormDialogProps, 'open'>) {
-  const [isPending, startTransition] = useTransition();
-
-  const availableTypes = alertTypes.filter(
-    (t) =>
-      !existingRuleTypes.includes(t.type) ||
-      (existingRule && existingRule.alert_type === t.type),
-  );
-
-  // All three seeded at mount and never re-seeded. The body only exists while
-  // the dialog is open (Base UI unmounts the portal once the close animation
-  // finishes), and the `key` above remounts it when the rule being edited
-  // changes, so there is nothing left for an adjustment effect to do.
-  const [routingMap, setRoutingMap] = useState<RoutingMap>(() =>
-    initialRoutingMap(existingRule, integrations),
-  );
-  const [routingErrors, setRoutingErrors] = useState<Record<number, string>>(
-    {},
-  );
-
-  const form = useForm<AlertRuleFormData>({
-    resolver: zodResolver(alertRuleFormSchema),
-    defaultValues: existingRule
-      ? {
-          name: existingRule.name,
-          alert_type: existingRule.alert_type,
-          selected_integration_ids: existingRule.integration_ids,
-          is_enabled: existingRule.is_enabled,
-          cooldown_minutes: existingRule.cooldown_minutes,
-        }
-      : {
-          name: '',
-          alert_type: availableTypes[0]?.type ?? 'new_issue',
-          selected_integration_ids: [],
-          is_enabled: true,
-          cooldown_minutes: 0,
-        },
-  });
-
-  const selectedIds = form.watch('selected_integration_ids');
-  const cooldown = form.watch('cooldown_minutes');
-  const alertType = form.watch('alert_type');
-
-  const toggleIntegration = (id: number, enabled: boolean) => {
-    const current = form.getValues('selected_integration_ids');
-    if (enabled) {
-      form.setValue('selected_integration_ids', [...current, id], {
-        shouldValidate: true,
-      });
-      form.clearErrors('selected_integration_ids');
-      setRoutingMap((prev) => ({ ...prev, [id]: {} }));
-    } else {
-      form.setValue(
-        'selected_integration_ids',
-        current.filter((i) => i !== id),
-      );
-      setRoutingMap((prev) => {
-        const next = { ...prev };
-        delete next[id];
-        return next;
-      });
-      setRoutingErrors((prev) => {
-        const next = { ...prev };
-        delete next[id];
-        return next;
-      });
-    }
-  };
-
-  const updateRoutingField = (
-    integrationId: number,
-    field: string,
-    value: string,
-  ) => {
-    setRoutingMap((prev) => ({
-      ...prev,
-      [integrationId]: { ...(prev[integrationId] ?? {}), [field]: value },
-    }));
-    setRoutingErrors((prev) => {
-      const next = { ...prev };
-      delete next[integrationId];
-      return next;
-    });
-  };
-
-  const onSubmit = (data: AlertRuleFormData) => {
-    const errors: Record<number, string> = {};
-    for (const id of data.selected_integration_ids) {
-      const integration = integrations.find((i) => i.id === id);
-      if (!integration) continue;
-      const routing = routingMap[id] ?? {};
-      const err = validateRoutingForIntegration(integration, routing);
-      if (err) errors[id] = err;
-    }
-
-    if (Object.keys(errors).length > 0) {
-      setRoutingErrors(errors);
-      return;
-    }
-
-    if (data.selected_integration_ids.length === 0) {
-      form.setError('selected_integration_ids', {
-        message: 'Select at least one integration',
-      });
-      return;
-    }
-
-    startTransition(async () => {
-      const channelsPayload = data.selected_integration_ids.flatMap((id) => {
-        const routing = routingMap[id] ?? {};
-        const integration = integrations.find((i) => i.id === id);
-        if (!integration) return [];
-        const routingOverride: Record<string, unknown> = {};
-
-        if (
-          integration.provider_type === 'slack' &&
-          getSlackMethod(integration) === 'bot_token'
-        ) {
-          if (routing.channel) routingOverride.channel = routing.channel.trim();
-        }
-        if (integration.provider_type === 'email' && routing.recipients) {
-          routingOverride.recipients = routing.recipients
-            .split(',')
-            .map((a) => a.trim())
-            .filter(Boolean);
-        }
-        if (integration.provider_type === 'webhook' && routing.url) {
-          routingOverride.url = routing.url.trim();
-        }
-
-        return [{ integration_id: id, routing_override: routingOverride }];
-      });
-
-      const result = existingRule
-        ? await updateAlertRule(projectId, existingRule.id, {
-            name: data.name,
-            is_enabled: data.is_enabled,
-            channels: channelsPayload,
-            cooldown_minutes: data.cooldown_minutes,
-          })
-        : await createAlertRule(projectId, {
-            name: data.name,
-            alert_type: data.alert_type,
-            channels: channelsPayload,
-            is_enabled: data.is_enabled,
-            cooldown_minutes: data.cooldown_minutes,
-          });
-
-      if (!result.success) {
-        // A duplicate rule name is a `conflict` naming `name`, and a second
-        // rule for an alert type already covered is one naming `alert_type`.
-        // Both are fields this dialog registers, so both land on their input.
-        const applied = applyServerFieldErrors(form, result.error, {
-          labels: { name: 'That name', alert_type: 'That alert type' },
-        });
-
-        if (applied.formLevel) {
-          toast.error('Failed to save rule', {
-            description: applied.formLevel,
-          });
-        }
-        return;
-      }
-
-      toast.success(existingRule ? 'Alert rule updated' : 'Alert rule created');
-      onSuccess();
-    });
-  };
-
-  return (
-    <>
-      <DialogHeader>
-        <DialogTitle>
-          {existingRule ? 'Edit Alert Rule' : 'New Alert Rule'}
-        </DialogTitle>
-        <DialogDescription>
-          {existingRule
-            ? 'Update when and where this rule sends notifications.'
-            : 'Define when to trigger an alert and where to send it.'}
-        </DialogDescription>
-      </DialogHeader>
-
-      <Form {...form}>
-        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-5 py-1">
-          {/* Name */}
-          <FormField
-            control={form.control}
-            name="name"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel className="text-xs font-bold uppercase tracking-widest text-muted-foreground">
-                  Rule Name
-                </FormLabel>
-                <FormControl>
-                  <Input
-                    placeholder="e.g., Notify team on new issues"
-                    autoComplete="off"
-                    disabled={isPending}
-                    {...field}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          {/* Trigger — radio cards */}
-          <FormField
-            control={form.control}
-            name="alert_type"
-            render={() => (
-              <FormItem>
-                <FormLabel className="text-xs font-bold uppercase tracking-widest text-muted-foreground">
-                  Trigger
-                </FormLabel>
-                <div className="grid grid-cols-3 gap-2 mt-1">
-                  {availableTypes.map((t) => {
-                    const Icon = t.icon;
-                    const isSelected = alertType === t.type;
-                    return (
-                      <button
-                        key={t.type}
-                        type="button"
-                        disabled={isPending || !!existingRule}
-                        onClick={() =>
-                          form.setValue('alert_type', t.type, {
-                            shouldValidate: true,
-                          })
-                        }
-                        className={cn(
-                          'flex flex-col items-start gap-1.5 rounded-lg border p-3 text-left transition-all',
-                          'hover:border-primary/50 hover:bg-accent/50',
-                          isSelected
-                            ? 'border-primary bg-primary/5 ring-1 ring-primary/20'
-                            : 'border-border bg-card',
-                          (isPending || !!existingRule) &&
-                            'opacity-60 cursor-not-allowed',
-                        )}
-                      >
-                        <Icon
-                          className={cn(
-                            'size-4',
-                            isSelected
-                              ? 'text-primary'
-                              : 'text-muted-foreground',
-                          )}
-                        />
-                        <div>
-                          <p
-                            className={cn(
-                              'text-xs font-semibold leading-none',
-                              isSelected ? 'text-primary' : 'text-foreground',
-                            )}
-                          >
-                            {t.name}
-                          </p>
-                          <p className="text-[10px] text-muted-foreground mt-0.5 leading-snug">
-                            {t.description}
-                          </p>
-                        </div>
-                      </button>
-                    );
-                  })}
-                </div>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          {/* Send To — integration toggle cards */}
-          <div>
-            <p className="text-xs font-bold uppercase tracking-widest text-muted-foreground mb-2">
-              Send To
-            </p>
-            <div className="space-y-2">
-              {integrations.map((integration) => {
-                const isSelected = selectedIds.includes(integration.id);
-                const routing = routingMap[integration.id] ?? {};
-                const routingErr = routingErrors[integration.id];
-                const isSlackBot =
-                  integration.provider_type === 'slack' &&
-                  getSlackMethod(integration) === 'bot_token';
-                const needsChannel = isSlackBot;
-                const needsRecipients = integration.provider_type === 'email';
-                const credUrl = (
-                  integration.credentials as Record<string, unknown>
-                ).url as string | undefined;
-                const needsUrl =
-                  integration.provider_type === 'webhook' && !credUrl;
-                const hasRoutingFields =
-                  needsChannel ||
-                  needsRecipients ||
-                  integration.provider_type === 'webhook';
-
-                return (
-                  <div key={integration.id}>
-                    {/* Toggle card — never changes background, only border */}
-                    <div
-                      className={cn(
-                        'flex items-center justify-between rounded-lg border bg-card px-3 py-2.5 transition-colors',
-                        isSelected ? 'border-primary/50' : 'border-border',
-                      )}
-                    >
-                      <div className="flex items-center gap-2.5">
-                        <div
-                          className={cn(
-                            'flex size-7 items-center justify-center rounded-md transition-colors',
-                            isSelected
-                              ? 'bg-primary/10 text-primary'
-                              : 'bg-muted text-muted-foreground',
-                          )}
-                        >
-                          <ProviderIcon
-                            type={integration.provider_type}
-                            className="size-3.5"
-                          />
-                        </div>
-                        <div>
-                          <p className="text-sm font-medium leading-none">
-                            {integration.name}
-                          </p>
-                          <p className="text-[10px] text-muted-foreground mt-0.5 capitalize">
-                            {integration.provider_type}
-                            {isSlackBot ? ' · bot token' : ''}
-                          </p>
-                        </div>
-                      </div>
-                      <Switch
-                        checked={isSelected}
-                        onCheckedChange={(checked) =>
-                          toggleIntegration(integration.id, checked)
-                        }
-                        disabled={isPending}
-                        size="sm"
-                      />
-                    </div>
-
-                    {/* Routing fields — separate block below the card */}
-                    {isSelected && hasRoutingFields && (
-                      <div className="mx-3 rounded-b-lg border border-t-0 border-primary/20 bg-muted/40 px-3 py-3 space-y-3">
-                        {needsChannel && (
-                          <div className="space-y-1.5">
-                            <label className="text-xs font-medium text-muted-foreground">
-                              Channel{' '}
-                              <span className="text-destructive">*</span>
-                            </label>
-                            <Input
-                              placeholder="#alerts or C1234567890"
-                              value={routing.channel ?? ''}
-                              onChange={(e) =>
-                                updateRoutingField(
-                                  integration.id,
-                                  'channel',
-                                  e.target.value,
-                                )
-                              }
-                              disabled={isPending}
-                              className="h-8 text-sm"
-                              autoComplete="off"
-                            />
-                            <p className="text-[10px] text-muted-foreground">
-                              Channel name (e.g.{' '}
-                              <code className="font-mono">#alerts</code>) or
-                              channel ID (e.g.{' '}
-                              <code className="font-mono">C1234567890</code>)
-                            </p>
-                          </div>
-                        )}
-                        {needsRecipients && (
-                          <div className="space-y-1.5">
-                            <label className="text-xs font-medium text-muted-foreground">
-                              Recipients{' '}
-                              <span className="text-destructive">*</span>
-                            </label>
-                            <Textarea
-                              placeholder="alerts@example.com, team@example.com"
-                              value={routing.recipients ?? ''}
-                              onChange={(e) =>
-                                updateRoutingField(
-                                  integration.id,
-                                  'recipients',
-                                  e.target.value,
-                                )
-                              }
-                              disabled={isPending}
-                              className="text-sm min-h-15 resize-none"
-                              autoComplete="off"
-                            />
-                            <p className="text-[10px] text-muted-foreground">
-                              Comma-separated email addresses
-                            </p>
-                          </div>
-                        )}
-                        {integration.provider_type === 'webhook' && (
-                          <div className="space-y-1.5">
-                            <label className="text-xs font-medium text-muted-foreground">
-                              Override URL
-                              {needsUrl ? (
-                                <span className="text-destructive"> *</span>
-                              ) : (
-                                <span className="text-muted-foreground/60">
-                                  {' '}
-                                  (optional)
-                                </span>
-                              )}
-                            </label>
-                            <Input
-                              placeholder="https://svc.io/hook"
-                              value={routing.url ?? ''}
-                              onChange={(e) =>
-                                updateRoutingField(
-                                  integration.id,
-                                  'url',
-                                  e.target.value,
-                                )
-                              }
-                              disabled={isPending}
-                              className="h-8 text-sm"
-                              autoComplete="off"
-                            />
-                            {!needsUrl && (
-                              <p className="text-[10px] text-muted-foreground">
-                                Overrides the URL configured in credentials
-                              </p>
-                            )}
-                          </div>
-                        )}
-                        {routingErr && (
-                          <p className="text-xs text-destructive">
-                            {routingErr}
-                          </p>
-                        )}
-                      </div>
-                    )}
-
-                    {isSelected && routingErr && !hasRoutingFields && (
-                      <p className="mx-3 text-xs text-destructive">
-                        {routingErr}
-                      </p>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
-            {form.formState.errors.selected_integration_ids && (
-              <p className="text-xs text-destructive mt-1.5">
-                {form.formState.errors.selected_integration_ids.message}
-              </p>
-            )}
-          </div>
-
-          {/* Cooldown */}
-          <FormField
-            control={form.control}
-            name="cooldown_minutes"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel className="text-xs font-bold uppercase tracking-widest text-muted-foreground">
-                  Cooldown
-                </FormLabel>
-                <div className="flex items-center gap-2">
-                  <Button
-                    type="button"
-                    variant="outline"
-                    size="icon"
-                    aria-label="Decrease cooldown by 5 minutes"
-                    className="size-9 shrink-0"
-                    disabled={isPending || cooldown <= 0}
-                    onClick={() => field.onChange(Math.max(0, cooldown - 5))}
-                  >
-                    <Minus className="size-3.5" />
-                  </Button>
-                  <FormControl>
-                    <Input
-                      type="number"
-                      min={0}
-                      disabled={isPending}
-                      className="text-center"
-                      {...field}
-                      onChange={(e) =>
-                        field.onChange(parseInt(e.target.value, 10) || 0)
-                      }
-                    />
-                  </FormControl>
-                  <Button
-                    type="button"
-                    variant="outline"
-                    size="icon"
-                    aria-label="Increase cooldown by 5 minutes"
-                    className="size-9 shrink-0"
-                    disabled={isPending}
-                    onClick={() => field.onChange(cooldown + 5)}
-                  >
-                    <Plus className="size-3.5" />
-                  </Button>
-                  <span className="text-sm text-muted-foreground shrink-0">
-                    min
-                  </span>
-                </div>
-                <FormDescription>
-                  {cooldown === 0
-                    ? 'No limit — alert fires every time'
-                    : `At most one alert every ${cooldown} minute${cooldown !== 1 ? 's' : ''} per issue`}
-                </FormDescription>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          {/* Enabled toggle */}
-          <FormField
-            control={form.control}
-            name="is_enabled"
-            render={({ field }) => (
-              <FormItem className="flex items-center justify-between rounded-lg border p-3">
-                <div>
-                  <FormLabel className="text-sm font-medium">
-                    Enable rule
-                  </FormLabel>
-                  <FormDescription className="text-xs">
-                    Start sending alerts immediately after saving
-                  </FormDescription>
-                </div>
-                <FormControl>
-                  <Switch
-                    checked={field.value}
-                    onCheckedChange={field.onChange}
-                    disabled={isPending}
-                  />
-                </FormControl>
-              </FormItem>
-            )}
-          />
-
-          <DialogFooter>
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => onOpenChange(false)}
-              disabled={isPending}
-            >
-              Cancel
-            </Button>
-            <Button type="submit" disabled={isPending}>
-              {isPending && <Loader2 className="size-4 mr-2 animate-spin" />}
-              {existingRule ? 'Save Changes' : 'Create Rule'}
-            </Button>
-          </DialogFooter>
-          {/* Where a failure that named no field of this form lands. */}
-          <FormRootError />
-        </form>
-      </Form>
     </>
   );
 }
